ml-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
